PriceTracker/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
beautifulsoup4==4.12.2
requests==2.31.0
APScheduler==3.10.4
python-dotenv==1.0.0
